% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_parent_locs}
\alias{picola_parent_locs}
\title{\emph{Pinus contorta} ssp. \emph{latifolia} parent tree locations}
\format{
A data frame with 2739 rows and 5 variables:
\describe{
\item{\code{Clone}}{integer Genotype identifier / Clone label for parent tree}
\item{\code{SPU}}{character Seed Planting Unit (SPU) identifier}
\item{\code{lat}}{double Latitude of parent tree}
\item{\code{lon}}{double Longitude of parent tree}
\item{\code{el}}{integer Elevation of parent tree (meters)}
}
}
\source{
The geographic data (latitude, longitude, and elevation) of each genotype's parent tree was obtained from the British Columbia Ministry of Forests Seed Planning and Registry Application (https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/tree-seed/seed-planning-use/spar).
}
\usage{
picola_parent_locs
}
\description{
Locations of genotype parents
}
\details{
Scions and seeds for \emph{Pinus contorta} ssp. \emph{latifolia} were collected from parent trees located in 143 natural stands across British Columbia during the 1970s and 1980s. These trees were selected based on their suitability for various breeding and selection programs that aimed to improve tree growth and adaptation to different B.C. regions. The scions collected from these trees were grafted onto rootstock to create a genotype bank and then from there onto rootstocks in seed orchards.
}
\keyword{datasets}
