% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_site_coord_elev}
\alias{picola_site_coord_elev}
\title{Site locations}
\format{
A data frame with 9 rows and 6 variables:
\describe{
\item{\code{Site}}{character Short symbolic name of the site}
\item{\code{orchard}}{logical Whether the site is a seed orchard (TRUE) or a comparison site (FALSE)}
\item{\code{longname}}{character Full descriptive name of the site}
\item{\code{lat}}{double Latitude of the site}
\item{\code{lon}}{double Longitude of the site}
\item{\code{el}}{double Elevation of the site (meters)}
}
}
\source{
Seed Orchard latitude and longitude data identified by C.S. Tysor manually using Google Earth, and verified by Jack Woods as the locations of these seed orchards. Elevation data for sites pulled from GeoNames via \code{rgbif::elevation}
}
\usage{
picola_site_coord_elev
}
\description{
Locations and elevations of \emph{Pinus contorta} ssp. \emph{latifolia} seed orchards and comparison sites
}
\keyword{datasets}
