% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_SPUs}
\alias{picola_SPUs}
\title{Seed orchard sites and SPUs}
\format{
A data frame with 33 rows and 9 variables:
\describe{
\item{\code{SPU_Number}}{integer SPU identifier}
\item{\code{SPU_Name}}{character Short symbolic name of the SPU}
\item{\code{SPU_Name_Long}}{character Full descriptive name of the SPU}
\item{\code{Long_Site}}{character Full descriptive name of the site}
\item{\code{Site}}{character Short symbolic name of the site}
\item{\code{Orchard}}{integer Orchard identifier within the site}
\item{\code{Repro_Data}}{integer Number of years of data for orchard at site}
\item{\code{Elev_Min}}{integer Minimum elevation of the SPU in meters}
\item{\code{Elev_Max}}{integer Maximum elevation of the SPU in meters}
}
}
\source{
https://www2.gov.bc.ca//gov/content/industry/forestry/managing-our-forest-resources/tree-seed/seed-planning-use/seed-planning-units-species-plans
}
\usage{
picola_SPUs
}
\description{
This dataset lists all the B.C. \emph{Pinus contorta} ssp. \emph{latifolia} seed orchards and the target SPUs for each orchard. Also includes the number of years of data in this package for each site-SPU combination.
}
\details{
This dataset lists all \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard sites and their corresponding Seed Planning Units (SPUs). Each entry includes the orchard identifier, the SPU number, and the elevation range for the SPU. Additionally, the dataset provides the number of years of data available for each site-SPU combination.
Seed Planning Units (SPUs) are defined regions used for organizing breeding and seed production efforts. These units are classified by species, seed planning zone, and elevation band.
}
\keyword{datasets}
